Import code from GPL archive received from QNAP

+++ include/linux/types.h
@@ -142,9 +142,9 @@ typedef __u16 __bitwise __sum16;
 typedef __u32 __bitwise __wsum;
 
 #ifdef CONFIG_PHYS_ADDR_T_64BIT
-typedef u64 phys_addr_t;
+typedef size_t phys_addr_t;
 #else
-typedef u32 phys_addr_t;
+typedef size_t phys_addr_t;
 #endif
 
 typedef phys_addr_t resource_size_t;
